增加狀態碼

+++ src/main/java/com/randy/masterspring/user/api/UserApiController.java
@@ -3,6 +3,8 @@ package com.randy.masterspring.user.api;
 import com.randy.masterspring.user.model.User;
 import com.randy.masterspring.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -20,17 +22,30 @@ public class UserApiController {
     }
 
     @RequestMapping(value = "/users",method = RequestMethod.POST)
-    public User create(@RequestBody User user){
-        return userRepository.save(user);
+    public ResponseEntity<User> create(@RequestBody User user){
+        HttpStatus status = HttpStatus.OK;
+        if (!userRepository.exists(user.getEmail())){
+            status = HttpStatus.CREATED;
+        }
+        User saved = userRepository.save(user);
+        return new ResponseEntity<>(saved,status);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.PUT)
-    public User updateUser(@PathVariable String email, @RequestBody User user){
-        return userRepository.save(email,user);
+    public ResponseEntity<User> updateUser(@PathVariable String email, @RequestBody User user){
+        if(!userRepository.exists(email)){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        User saved = userRepository.save(email,user);
+        return new ResponseEntity<User>(saved,HttpStatus.CREATED);
     }
 
     @RequestMapping(value = "/user/{email}", method = RequestMethod.DELETE)
-    public void deleteUser(@PathVariable String email){
+    public ResponseEntity<User> deleteUser(@PathVariable String email){
+        if (!userRepository.exists(email)){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         userRepository.delete(email);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
